+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added `fillto` to `BarPlot` aesthetics [#535](https://github.com/MakieOrg/AlgebraOfGraphics.jl/pull/535).
+- Fixed bug when giving `datalimits` of `density` as a (low, high) tuple [#536](https://github.com/MakieOrg/AlgebraOfGraphics.jl/pull/536).
 
 ## v0.8.3 - 2024-08-23